requirements.txt
streamlit==1.31.1
pandas
plotly
scipy==1.11.4
scikit-learn
pyyaml